Create AuthenticationSessionManager to handle authentication
