+++ Sources/Classes/Data/Repositories/AuthenticationRepository.swift
@@ -19,16 +19,16 @@ final class AuthenticationRepository: AuthenticationRepositoryProtocol {
     }
     
     func getToken(code: String, _ completion: @escaping RequestCompletion<APIToken>) {
-        guard let clientId: String = try? keychain.get(.clientId),
-              let clientSecret: String = try? keychain.get(.clientSecret) else {
-            completion(.failure(.init(error: "Missing client id or client secret")))
+        guard let clientId: String = try? keychain.get(.clientId) else {
+            completion(.failure(.init(error: "Missing client id")))
             return
         }
+        let clientSecret: String? = try? keychain.get(.clientSecret)
         networkAPI.performRequest(
             .getToken(
                 code: code,
                 clientId: clientId,
-                clientSecret: clientSecret
+                clientSecret: clientSecret ?? ""
             ),
             completion: completion
         )
